Change popularity type from Double to Float in CastMember

+++ core/domain/src/main/kotlin/com/koniukhov/cinecircle/core/domain/model/CastMember.kt
@@ -7,7 +7,7 @@ data class CastMember(
     val knownForDepartment: String,
     val name: String,
     val originalName: String,
-    val popularity: Double,
+    val popularity: Float,
     val profilePath: String,
     val castId: Int,
     val character: String,
